Add experimental async PassAsync method to skUnit
Added an experimental asynchronous method `PassAsync` to the `skUnit` namespace in the `ScenarioAssert_ChatClient.cs` file. This method, marked with the `[Experimental("SKEXP0001")]` attribute, takes a `ChatScenario` object, a `Kernel` object, and an optional list of `ChatMessage` objects as parameters. It enhances the functionality of the `skUnit` namespace by providing an asynchronous way to pass chat scenarios to a kernel, potentially improving performance and responsiveness.